added collate function to handle meshes of different sizes. It slightly slow-down the training loop (~5%). I don't know the impact on the validation with respect to reading the mesh from disk and transforming it
Now the model is started from a specific configuration file e.g. if we want to execute the server configuration we have to run server_config.py. This makes more explicit which configuration we are using.

if we want to run a file without passing from a configuration we'll have to modify configs/__init__.py so that it selects the one we want.